refactor: rename datastore
