+++ http/static/js/api/datastores.js
