+++ http/static/js/widget/datastores/create.js
