+++ http/static/js/widget/datastores/iscsi/create.js
