+++ http/static/js/widget/datastores/nfs/create.js
