+++ http/static/js/widget/datastores/upload.js
